creating new component
